Fix add_read_acl API & impl, update token acquisition - see below
add_read_acl should not explicitly take a token as an argument. It's
passed in the header.

add_read_acl should only work for specific users defined in deploy.cfg.

Rework token acquisition to avoid system calls.

Some minor tweaks/fixes to the acl updating code - this is untested for
now, because:

TODO - need to make a call on whether the handle manager contacts the
hsi database directly or has admin rights for the handle_id_to_handle
method on the Handle Service.