Hacky workaround for auth read timeouts
For some starman workers I was seeing a read timeout even with
a fairly conservative timeout set in AuthToken (10 seconds).  It
is not clear where this problem is happening, but my wild guess
is that the current auth service is not able to keep up with that
many simultaneous requests.  (The handle manager starts 20 workers,
which seems like a lot, but if the auth service were also a starman
with only a few workers I could imagine this being an issue.  The
current auth service is not starman but it could still have a similar
issue.)

At this point it is not worth fixing the current auth service, or
honestly even trying to troubleshoot the issue.  So I've added a
very crude stopgap measure: each worker will sleep for a random
number of seconds before trying to talk to the auth service.

When the new auth service is ready, we should remove this line
and re-test against the new service.  If there are still 500 read
errors from LWP, perhaps the issue is somewhere else, but in any
case that would be the point at which we should spend cycles on
the problem.